requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Authlib==1.6.0
autoflake==2.3.1
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coverage==7.8.0
cryptography==45.0.3
Django==5.2
django-axes==8.0.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dparse==0.6.4
drf-yasg==1.21.10
filelock==3.16.1
flake8==7.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nltk==3.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
psutil==6.1.1
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.13
ruamel.yaml.clib==0.2.12
safety==3.5.2
safety-schemas==0.0.14
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing_extensions==4.14.0
uritemplate==4.1.1
urllib3==2.4.0
